% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generate_script_of_multiple_profileOfSpeciesHTMLs.R
\name{generate_script_of_multiple_profileOfSpeciesHTMLs}
\alias{generate_script_of_multiple_profileOfSpeciesHTMLs}
\title{Generate the script to produce multiple HTMLs of the panel of species profiles}
\usage{
generate_script_of_multiple_profileOfSpeciesHTMLs(flow = c("PA", "PNA"))
}
\description{
Gera o script para produzir múltiplos painéis dos perfis das espécies
}
