% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLength.R
\name{genLength}
\alias{genLength}
\title{Generation length}
\usage{
genLength(firstRep, lastRep)
}
\arguments{
\item{firstRep}{First year of reproduction (numeric)}

\item{lastRep}{Last year of reproduction (numeric)}
}
\value{
a numeric value
}
\description{
Calcula o tempo de geração
}
\author{
Eduardo Amorim, \email{eduardoamorim@jbrj.gov.br}
}
