% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_occurrenceRecords_from_oldSystem.R
\name{get_occurrenceRecords_from_oldSystem}
\alias{get_occurrenceRecords_from_oldSystem}
\title{Load occurrence records from old system of CNCFlora}
\usage{
get_occurrenceRecords_from_oldSystem(Species)
}
\description{
Carrega os dados dos registros de ocorrências do arquivo HTML obtido a partir do antigo sistema CNCFlora
}
