% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlayAnalysis_landCoverMapBiomas.R
\name{overlayAnalysis_landCoverMapBiomas}
\alias{overlayAnalysis_landCoverMapBiomas}
\title{Overlay Analysis between occurrence records and historical series of cover land from MapBiomas}
\usage{
overlayAnalysis_landCoverMapBiomas()
}
\description{
Análise de sobreposição entre os registros de ocorrências e a série histórica de cobertura do solo do MapBiomas
}
