% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundDownFirstDecimalPlace.R
\name{roundDownFirstDecimalPlace}
\alias{roundDownFirstDecimalPlace}
\title{Round down to one decimal place}
\usage{
roundDownFirstDecimalPlace(x)
}
\description{
Arredonda a primeira casa decimal para baixo
}
