% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundUpFirstDecimalPlace.R
\name{roundUpFirstDecimalPlace}
\alias{roundUpFirstDecimalPlace}
\title{Round up to one decimal place}
\usage{
roundUpFirstDecimalPlace(x)
}
\description{
Arredonda a primeira casa decimal para cima
}
