% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/scrap_data_of_filledSpeciesProfile_from_oldSystem.R
\name{scrap_data_of_filledSpeciesProfile_from_oldSystem}
\alias{scrap_data_of_filledSpeciesProfile_from_oldSystem}
\title{Scraping of data on filled profile of species from CNCFlora old System}
\usage{
scrap_data_of_filledSpeciesProfile_from_oldSystem(filledSpeciesProfile)
}
\description{
Obtém os dados dos perfis das espécies preenchidos do sistema antigo do CNCFlora
}
