% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationCoordinates.R
\name{validationCoordinates}
\alias{validationCoordinates}
\title{Verify the validation of Coordinates and fix errors}
\usage{
validationCoordinates()
}
\description{
Verifica se há erros tipográficos nas coordenadas e os corrigede
}
