% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validationOccurrences.R
\name{validationOccurrences}
\alias{validationOccurrences}
\title{Verify the validation of occurrence records}
\usage{
validationOccurrences()
}
\description{
Verifica se o arquivo corresponde ao nome da espécie e se há registros não validados
}
